+++ cpp/src/lib/messages/cucumber/messages/pickle.cpp
@@ -12,6 +12,7 @@ pickle::to_string() const
 
     cucumber::messages::to_string(oss, "id=", id);
     cucumber::messages::to_string(oss, ", uri=", uri);
+    cucumber::messages::to_string(oss, ", location=", location);
     cucumber::messages::to_string(oss, ", name=", name);
     cucumber::messages::to_string(oss, ", language=", language);
     cucumber::messages::to_string(oss, ", steps=", steps);
@@ -26,6 +27,7 @@ pickle::to_json(json& j) const
 {
     cucumber::messages::to_json(j, camelize("id"), id);
     cucumber::messages::to_json(j, camelize("uri"), uri);
+    cucumber::messages::to_json(j, camelize("location"), location);
     cucumber::messages::to_json(j, camelize("name"), name);
     cucumber::messages::to_json(j, camelize("language"), language);
     cucumber::messages::to_json(j, camelize("steps"), steps);