+++ go/messages.go
@@ -218,6 +218,7 @@ type ParseError struct {
 type Pickle struct {
 	Id         string        `json:"id"`
 	Uri        string        `json:"uri"`
+	Location   *Location     `json:"location,omitempty"`
 	Name       string        `json:"name"`
 	Language   string        `json:"language"`
 	Steps      []*PickleStep `json:"steps"`